requirements.txt
Requirements
Python 3.8

discord.py

psutil